+++ R/dataframe__frame.R
@@ -1935,8 +1935,7 @@ DataFrame_transpose = function(
 #' * `"never"`: This never puts quotes around fields, even if that results in
 #'   invalid CSV data (e.g. by not quoting strings containing the separator).
 #'
-#' @return
-#' This doesn't return anything.
+#' @return Invisibly returns the input DataFrame.
 #'
 #' @rdname IO_write_csv
 #'
@@ -1970,7 +1969,7 @@ DataFrame_write_csv = function(
   ) |>
     unwrap("in $write_csv():")
 
-  invisible(NULL)
+  invisible(self)
 }
 
 
@@ -2009,7 +2008,7 @@ DataFrame_write_ipc = function(
   ) |>
     unwrap("in $write_ipc():")
 
-  invisible(NULL)
+  invisible(self)
 }
 
 
@@ -2044,7 +2043,7 @@ DataFrame_write_parquet = function(
   ) |>
     unwrap("in $write_parquet():")
 
-  invisible(NULL)
+  invisible(self)
 }
 
 #' Write to JSON file
@@ -2075,7 +2074,7 @@ DataFrame_write_json = function(
   .pr$DataFrame$write_json(self, file, pretty, row_oriented) |>
     unwrap("in $write_json():")
 
-  invisible(NULL)
+  invisible(self)
 }
 
 #' Write to NDJSON file
@@ -2096,7 +2095,7 @@ DataFrame_write_ndjson = function(file) {
   .pr$DataFrame$write_ndjson(self, file) |>
     unwrap("in $write_ndjson():")
 
-  invisible(NULL)
+  invisible(self)
 }
 
 #' @inherit LazyFrame_rolling title description params details